+++ backend/src/baserow/api/exceptions.py
@@ -33,3 +33,12 @@ class InvalidClientSessionIdAPIException(APIException):
         f"be between 1 and {settings.MAX_CLIENT_SESSION_ID_LENGTH} characters long and "
         f"must only contain alphanumeric or the - characters.",
     )
+
+
+class InvalidUndoRedoActionGroupIdAPIException(APIException):
+    status_code = status.HTTP_400_BAD_REQUEST
+    default_code = "ERROR_INVALID_CLIENT_ACTION_GROUP"
+    default_detail = (
+        f"An invalid {settings.CLIENT_UNDO_REDO_ACTION_GROUP_ID_HEADER} header was provided. "
+        f"It must be a valid Version 4 UUID.",
+    )