+++ backend/src/baserow/test_utils/fixtures/user.py
@@ -3,7 +3,10 @@ from django.contrib.auth import get_user_model
 from rest_framework_jwt.settings import api_settings
 
 from baserow.core.models import UserProfile
-from baserow.api.sessions import set_untrusted_client_session_id
+from baserow.api.sessions import (
+    set_client_undo_redo_action_group_id,
+    set_untrusted_client_session_id,
+)
 
 User = get_user_model()
 jwt_payload_handler = api_settings.JWT_PAYLOAD_HANDLER
@@ -32,6 +35,7 @@ class UserFixtures:
             kwargs["password"] = "password"
 
         session_id = kwargs.pop("session_id", "default-test-user-session-id")
+        action_group = kwargs.pop("action_group", None)
 
         profile_data["language"] = kwargs.pop("language", "en")
         profile_data["to_be_deleted"] = kwargs.pop("to_be_deleted", False)
@@ -45,6 +49,7 @@ class UserFixtures:
         UserProfile.objects.create(**profile_data)
 
         set_untrusted_client_session_id(user, session_id)
+        set_client_undo_redo_action_group_id(user, action_group)
 
         return user