+++ backend/tests/baserow/core/actions/test_application_actions.py
@@ -13,6 +13,7 @@ from baserow.core.models import Application
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_redo_order_applications(data_fixture, django_assert_num_queries):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)
@@ -43,6 +44,7 @@ def test_can_undo_redo_order_applications(data_fixture, django_assert_num_querie
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_creating_application(data_fixture, django_assert_num_queries):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)
@@ -62,6 +64,7 @@ def test_can_undo_creating_application(data_fixture, django_assert_num_queries):
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_redo_creating_application(data_fixture, django_assert_num_queries):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)
@@ -84,6 +87,7 @@ def test_can_undo_redo_creating_application(data_fixture, django_assert_num_quer
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_deleteing_application(data_fixture):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)
@@ -107,6 +111,7 @@ def test_can_undo_deleteing_application(data_fixture):
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_redo_deleting_application(data_fixture, django_assert_num_queries):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)
@@ -134,6 +139,7 @@ def test_can_undo_redo_deleting_application(data_fixture, django_assert_num_quer
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_updating_application(data_fixture, django_assert_num_queries):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)
@@ -160,6 +166,7 @@ def test_can_undo_updating_application(data_fixture, django_assert_num_queries):
 
 
 @pytest.mark.django_db
+@pytest.mark.undo_redo
 def test_can_undo_redo_updating_application(data_fixture, django_assert_num_queries):
     session_id = "session-id"
     user = data_fixture.create_user(session_id=session_id)